requirements.txt
fleep==1.0.1
Pillow==7.1.2
pyinotify==0.9.6
